scrms/requirements.txt
asn1crypto==0.24.0
cffi==1.11.4
cryptography==2.1.4
Django==2.0.1
idna==2.6
Pillow==5.0.0
pycparser==2.18
pytz==2017.3
six==1.11.0
sorl-thumbnail==12.4.1
